# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pro Disciplinas",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/preOffer": {
            "get": {
                "tags": [
                    "/preOffer"
                ],
                "summary": "Display a listing of the resource",
                "description": "Get all pre-offer on database",
                "operationId": "1751fba8b3184ae0f5fd22df3643f9cf",
                "responses": {
                    "200": {
                        "description": "List of pre-offer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/preOffer"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Post pre-offer on database",
                "operationId": "57d1aeca7d3d7e6d1452a5baa2879796",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "shift": {
                                        "type": "string"
                                    },
                                    "formatType": {
                                        "type": "string"
                                    },
                                    "turma_id": {
                                        "type": "integer"
                                    },
                                    "period_id": {
                                        "type": "integer"
                                    },
                                    "discipline_id": {
                                        "type": "integer"
                                    },
                                    "siap_coordenador": {
                                        "type": "string"
                                    },
                                    "this_is_pro_discipline": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of pre-offer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/preOffer/{id}": {
            "get": {
                "tags": [
                    "/preOffer"
                ],
                "summary": "Display a specified resource",
                "description": "Show pre-offer",
                "operationId": "8da51e3426d9c83942a5040ccee9e39c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PreOffer id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show pre-offer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/preOffer"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update pre-offer on database",
                "operationId": "2c630636ed6cee6bfbf151921b484013",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PreOffer id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "shift": {
                                        "type": "string"
                                    },
                                    "formatType": {
                                        "type": "string"
                                    },
                                    "turma_id": {
                                        "type": "integer"
                                    },
                                    "period_id": {
                                        "type": "integer"
                                    },
                                    "siap_coordenador": {
                                        "type": "string"
                                    },
                                    "this_is_pro_disciplina": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update of pre-offer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/preOffer"
                ],
                "summary": " Remove the specified resource from storage.",
                "description": "Remove a pre-offer on database",
                "operationId": "c68d1151355c47b6b12d6ac4777122cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PreOffer Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "pre-offer deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/turma": {
            "get": {
                "tags": [
                    "/turma"
                ],
                "summary": "Display a listing of the resource",
                "description": "Get all turma on database",
                "operationId": "e89afe2a1d54444861986388893161f1",
                "responses": {
                    "200": {
                        "description": "List of turma"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "/turma"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Post turma on database",
                "operationId": "26fb32b15bfe1687b04f482f1bc9455d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "admission_year": {
                                        "type": "integer"
                                    },
                                    "course_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of turma"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/turma/{id}": {
            "get": {
                "tags": [
                    "/turma"
                ],
                "summary": "Display a specified resource",
                "description": "Show turma",
                "operationId": "ca6db55367ce0e761dfa48e0926d2a77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Turma id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show turma"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/turma"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update turma on database",
                "operationId": "e40f3f8d8bc802e6623415b9ea0532b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Turma id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "admission_year": {
                                        "type": "integer"
                                    },
                                    "course_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update of turma"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/turma"
                ],
                "summary": " Remove the specified resource from storage.",
                "description": "Remove a turma on database",
                "operationId": "9cb22f34c6b9d4d70b54b9cdcbb5c2ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Turma Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "turma deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}